fixed msvc warnings in audio helpers

+++ include/mostly_harmless/audio/mostlyharmless_AudioHelpers.h
@@ -44,8 +44,8 @@ namespace mostly_harmless {
         size_t lastEventIndex{ 0 };
         auto* const* rawBuff = buffer.getArrayOfWritePointers();
         for (auto i = 0; i < buffer.getNumSamples(); ++i) {
-            if (eventContext.next() && eventContext.next()->time == i) {
-                while (eventContext.next() && eventContext.next()->time == i) {
+            if (eventContext.next() && eventContext.next()->time == static_cast<std::uint32_t>(i)) {
+                while (eventContext.next() && eventContext.next()->time == static_cast<std::uint32_t>(i)) {
                     eventCallback(eventContext.next());
                     ++eventContext;
                 }
